requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0
groq>=0.4.0
tavily-python>=0.3.0
pymupdf>=1.23.0
python-dotenv>=1.0.0
pydantic>=2.0.0
